+++ src/com/intellij/idea/plugin/hybris/tools/ccv2/api/CCv2Api.kt
@@ -115,7 +115,7 @@ class CCv2Api {
         ?.map { CCv2DataBackupDto.map(it) }
         ?: emptyList()
 
-    suspend fun fetchEnvironmentsBuilds(
+    fun fetchEnvironmentsBuilds(
         ccv2Token: String,
         subscription: CCv2Subscription,
         environments: Collection<CCv2EnvironmentDto>,