+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Added support of the `metatype` for Item Types [#1345](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1345)
 - Added support of the `metatype` for Relation Elements [#1347](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1347)
 - Addressed deadlock for circular extends of two Item Types [#1348](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1348)
+- Reworked caching mechanism for Global TS Meta Model [#1351](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1351)
 
 ### `items.xml` inspection rules
 - Item Type `metatype` must be a type extending ComposedType [#1346](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1346)